Split host definition into its own file
