+++ component/component.go
@@ -58,43 +58,6 @@ const (
 	KindExtension
 )
 
-// Host represents the entity that is hosting a Component. It is used to allow communication
-// between the Component and its host (normally the service.Application is the host).
-type Host interface {
-	// ReportFatalError is used to report to the host that the extension
-	// encountered a fatal error (i.e.: an error that the instance can't recover
-	// from) after its start function had already returned.
-	ReportFatalError(err error)
-
-	// GetFactory of the specified kind. Returns the factory for a component type.
-	// This allows components to create other components. For example:
-	//   func (r MyReceiver) Start(host component.Host) error {
-	//     apacheFactory := host.GetFactory(KindReceiver,"apache").(component.ReceiverFactory)
-	//     receiver, err := apacheFactory.CreateMetricsReceiver(...)
-	//     ...
-	//   }
-	// GetFactory can be called by the component anytime after Start() begins and
-	// until Shutdown() is called. Note that the component is responsible for destroying
-	// other components that it creates.
-	GetFactory(kind Kind, componentType configmodels.Type) Factory
-
-	// Return map of extensions. Only enabled and created extensions will be returned.
-	// Typically is used to find an extension by type or by full config name. Both cases
-	// can be done by iterating the returned map. There are typically very few extensions
-	// so there there is no performance implications due to iteration.
-	GetExtensions() map[configmodels.NamedEntity]Extension
-
-	// Return map of exporters. Only enabled and created exporters will be returned.
-	// Typically is used to find exporters by type or by full config name. Both cases
-	// can be done by iterating the returned map. There are typically very few exporters
-	// so there there is no performance implications due to iteration.
-	// This returns a map by DataType of maps by exporter configs to the exporter instance.
-	// Note that an exporter with the same name may be attached to multiple pipelines and
-	// thus we may have an instance of the exporter for multiple data types.
-	// This is an experimental function that may change or even be removed completely.
-	GetExporters() map[configmodels.DataType]map[configmodels.NamedEntity]Exporter
-}
-
 // Factory interface must be implemented by all component factories.
 type Factory interface {
 	// Type gets the type of the component created by this factory.
